+++ browser/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0-beta.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/browser@5.0.0-beta.0...@capacitor/browser@5.0.0-beta.1) (2023-04-21)
+
+
+### Features
+
+* Update gradle to 8.0.2 and gradle plugin to 8.0.0 ([#1542](https://github.com/ionic-team/capacitor-plugins/issues/1542)) ([e7210b4](https://github.com/ionic-team/capacitor-plugins/commit/e7210b47867644f5983e37acdbf0247214ec232d))
+
+
+
+
+
+# [5.0.0-beta.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/browser@5.0.0-alpha.1...@capacitor/browser@5.0.0-beta.0) (2023-03-31)
+
+**Note:** Version bump only for package @capacitor/browser
+
+
+
+
+
+# [5.0.0-alpha.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/browser@4.1.1...@capacitor/browser@5.0.0-alpha.1) (2023-03-16)
+
+
+### Features
+
+* **android:** Removing enableJetifier ([d66f9cb](https://github.com/ionic-team/capacitor-plugins/commit/d66f9cbd9da7e3b1d8c64ca6a5b45156867d4a04))
+
+
+
+
+
+## [4.1.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/browser@4.1.0...@capacitor/browser@4.1.1) (2023-01-17)
+
+
+### Bug Fixes
+
+* **browser:** Reject if custom tabs are not available ([#1323](https://github.com/ionic-team/capacitor-plugins/issues/1323)) ([3c90fea](https://github.com/ionic-team/capacitor-plugins/commit/3c90fea63e2c24deacc38ba9fd8471c2cc0bde56))
+
+
+
+
+
 # [4.1.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/browser@1.0.7...@capacitor/browser@4.1.0) (2022-11-16)